requirements.txt
# Russian Learner Bot - Requirements
# Python 3.9+

# Telegram Bot Framework
aiogram==3.3.0

# AI & NLP
openai==1.6.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Scheduling
apscheduler==3.10.4

# Audio Processing
gtts==2.5.0
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Caching (optional but recommended)
redis==5.0.1
aioredis==2.0.1

# Monitoring & Logging
python-telegram-bot-logging==0.1.0
sentry-sdk==1.39.2

# Testing (development only)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0